Week 1

+++ Algorithmic Toolbox/Week 1 (Programming Challenges)/Maximum Bitwise Product/maximumbitwiseproduct.cpp
@@ -5,13 +5,13 @@ using std::cin;
 using std::cout;
 using std::vector;
 
-int MaxBitwiseProduct(const vector<int>& numbers) {
-    int result = 0;
+long long MaxBitwiseProduct(const vector<int>& numbers) {
+    long long result = 0;
     int n = numbers.size();
     for(int i = 0; i < n; i++){
         for(int j = i+1; j < n; j++){
-            if(numbers[i]*numbers[j]>result){
-                result = numbers[i]*numbers[j];
+            if(((long long )numbers[i])*numbers[j]>result){
+                result = ((long long )numbers[i])*numbers[j];
             }
         }
     }
@@ -24,7 +24,7 @@ int main(){
     for (int i = 0; i < n; ++i) {
         cin>>numbers[i];
     }
-    int result = MaxBitwiseProduct(numbers);
+    long long result = MaxBitwiseProduct(numbers);
     cout<<result<<"\n";
     return 0;
 }
